contract: move anchor module from validation
