+++ src/contract/seal.rs
@@ -21,7 +21,6 @@
 // limitations under the License.
 
 use core::fmt::Debug;
-use std::cmp::Ordering;
 use std::hash::Hash;
 use std::num::NonZeroU32;
 
@@ -187,59 +186,6 @@ impl WitnessOrd {
         WitnessHeight::new(height)
             .map(WitnessOrd::OnChain)
             .unwrap_or(WitnessOrd::OffChain)
-    }
-}
-
-/// Txid and height information ordered according to the RGB consensus rules.
-#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug, Display)]
-#[derive(StrictType, StrictDumb, StrictEncode, StrictDecode)]
-#[strict_type(lib = LIB_NAME_RGB, tags = custom, dumb = Self::Bitcoin(strict_dumb!(), strict_dumb!()))]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(crate = "serde_crate", rename_all = "camelCase")
-)]
-#[non_exhaustive]
-pub enum WitnessAnchor {
-    #[strict_type(tag = 0x00, dumb)]
-    #[display("bitcoin:{0}/{1}")]
-    Bitcoin(WitnessOrd, Txid),
-
-    #[strict_type(tag = 0x01)]
-    #[display("liquid:{0}/{1}")]
-    Liquid(WitnessOrd, Txid),
-}
-
-impl PartialOrd for WitnessAnchor {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> { Some(self.cmp(other)) }
-}
-
-impl Ord for WitnessAnchor {
-    fn cmp(&self, other: &Self) -> Ordering {
-        if self == other {
-            return Ordering::Equal;
-        }
-        match (self, other) {
-            (WitnessAnchor::Bitcoin(..), WitnessAnchor::Liquid(..)) => Ordering::Less,
-            (WitnessAnchor::Liquid(..), WitnessAnchor::Bitcoin(..)) => Ordering::Greater,
-            (
-                WitnessAnchor::Bitcoin(ord1, txid1) | WitnessAnchor::Liquid(ord1, txid1),
-                WitnessAnchor::Bitcoin(ord2, txid2) | WitnessAnchor::Liquid(ord2, txid2),
-            ) if ord1 == ord2 => txid1.cmp(txid2),
-            (
-                WitnessAnchor::Bitcoin(ord1, _) | WitnessAnchor::Liquid(ord1, _),
-                WitnessAnchor::Bitcoin(ord2, _) | WitnessAnchor::Liquid(ord2, _),
-            ) => ord1.cmp(ord2),
-        }
-    }
-}
-
-impl WitnessAnchor {
-    pub fn witness_id(self) -> WitnessId {
-        match self {
-            WitnessAnchor::Bitcoin(_, txid) => WitnessId::Bitcoin(txid),
-            WitnessAnchor::Liquid(_, txid) => WitnessId::Liquid(txid),
-        }
     }
 }
