+++ src/lib.rs
@@ -55,7 +55,6 @@ pub mod prelude {
     pub use bp::dbc::AnchorId;
     pub use contract::*;
     pub use schema::*;
-    pub use validation::{Anchor, AnchoredBundle, Layer1};
 
     use super::*;
     pub use super::{schema, vm};
